warnings: remove some compile-time warnings
There are still maaaaany more warnings.
Fixing them... might take a while

+++ src/charsetwidget.h
@@ -59,7 +59,7 @@ protected:
     QSize _selectingSize;
     int _charIndex;
     QSize _sizeHint;
-    float _zoomLevel;
+    qreal _zoomLevel;
     bool _displayGrid;
 };
